Fixed SLg Client/Server session implementation for received LRR/LRA in IDLE case, which made remaining SLg test pass for deferred location services. Issues #37 and #38, PR #111
